Remove handleRequest from react-sdk
We should not have any server specific functions in react-sdk.
Any small change can break client environment.

And moving to /server entry point still problematic because this
entry point is node specific while we want to support as many
environments as possible.

Recently remix added defaults for different environments so user
do not need to add any entry.client.tsx or entry.server.ts at all.

Defaults do not work for us in builder though doing great
in size test setup.
