+++ apps/builder/app/entry.server.tsx
@@ -1,14 +1,33 @@
+import { renderToString } from "react-dom/server";
 import type { EntryContext } from "@remix-run/node";
+import { RemixServer } from "@remix-run/react";
 import * as Sentry from "@sentry/remix";
-import { initSentry } from "./shared/sentry";
 import { prisma } from "@webstudio-is/prisma-client";
+import { initSentry } from "./shared/sentry";
 import { handleRequest as handleRequestBuilder } from "./shared/remix";
-import { handleRequest as handleRequestCanvas } from "@webstudio-is/react-sdk";
 import { getBuildParams } from "./shared/router-utils";
 
 initSentry({
   integrations: [new Sentry.Integrations.Prisma({ client: prisma })],
 });
+
+const handleRequestCanvas = (
+  request: Request,
+  responseStatusCode: number,
+  responseHeaders: Headers,
+  remixContext: EntryContext
+) => {
+  const markup = renderToString(
+    <RemixServer context={remixContext} url={request.url} />
+  );
+
+  responseHeaders.set("Content-Type", "text/html");
+
+  return new Response(`<!DOCTYPE html>${markup}`, {
+    status: responseStatusCode,
+    headers: responseHeaders,
+  });
+};
 
 export default function handleRequest(
   request: Request,
