+++ packages/react-sdk/package.json
@@ -20,11 +20,8 @@
   "devDependencies": {
     "@babel/core": "^7.21.0",
     "@jest/globals": "^29.3.1",
-    "@remix-run/node": "1.15.0",
-    "@remix-run/react": "1.15.0",
-    "@remix-run/server-runtime": "1.15.0",
+    "@remix-run/react": "^1.15.0",
     "@storybook/react": "^6.5.16",
-    "@types/node": "^18.11.18",
     "@types/react": "^18.0.35",
     "@types/react-dom": "^18.0.11",
     "@webstudio-is/jest-config": "workspace:^",
@@ -38,8 +35,7 @@
     "zod": "^3.19.1"
   },
   "peerDependencies": {
-    "@remix-run/react": "1.9.0",
-    "@remix-run/server-runtime": "1.9.0",
+    "@remix-run/react": "^1.15.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "zod": "^3.19.1"
