+++ packages/react-sdk/src/app/index.ts
@@ -1,3 +1,2 @@
 export * from "./root";
-export * from "./handle-request.server";
 export * from "./params";
