--- packages/sdk-size-test/app/entry.server.tsx
@@ -1,16 +1,0 @@
-import type { EntryContext } from "@remix-run/node";
-import { handleRequest as handleRequestRemix } from "@webstudio-is/react-sdk";
-
-export default function handleRequest(
-  request: Request,
-  responseStatusCode: number,
-  responseHeaders: Headers,
-  remixContext: EntryContext
-) {
-  return handleRequestRemix(
-    request,
-    responseStatusCode,
-    responseHeaders,
-    remixContext
-  );
-}
