+++ packages/sdk-size-test/package.json
@@ -20,16 +20,18 @@
     }
   ],
   "dependencies": {
-    "@remix-run/cloudflare-pages": "1.15.0",
-    "@remix-run/react": "1.15.0",
+    "@remix-run/cloudflare": "^1.15.0",
+    "@remix-run/cloudflare-pages": "^1.15.0",
+    "@remix-run/react": "^1.15.0",
     "@webstudio-is/project": "workspace:^",
     "@webstudio-is/project-build": "workspace:^",
     "@webstudio-is/react-sdk": "workspace:^",
+    "isbot": "^3.6.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "1.15.0",
+    "@remix-run/dev": "^1.15.0",
     "@size-limit/file": "^8.2.4",
     "@types/react": "^18.0.35",
     "@types/react-dom": "^18.0.11",
